[core] (3/n) [Removing GCS Centralized Scheduling] Removing the GCS's ability to send global GC requests to all raylets in the cluster. Moving the GC throttler to the Raylet's package now that it's only used by Raylets.
