+++ src/ray/raylet/BUILD.bazel
@@ -222,6 +222,15 @@ ray_cc_library(
 )
 
 ray_cc_library(
+    name = "throttler",
+    hdrs = ["throttler.h"],
+    visibility = [":__subpackages__"],
+    deps = [
+        "@com_google_absl//absl/time",
+    ],
+)
+
+ray_cc_library(
     name = "raylet_lib",
     srcs = [
         "node_manager.cc",
@@ -244,6 +253,7 @@ ray_cc_library(
         ":local_object_manager_interface",
         ":placement_group_resource_manager",
         ":runtime_env_agent_client",
+        ":throttler",
         ":wait_manager",
         ":worker",
         ":worker_killing_policy",
@@ -271,7 +281,6 @@ ray_cc_library(
         "//src/ray/util:container_util",
         "//src/ray/util:network_util",
         "//src/ray/util:port_persistence",
-        "//src/ray/util:throttler",
         "//src/ray/util:time",
         "@boost//:system",
         "@com_google_absl//absl/base:core_headers",
