+++ src/ray/raylet/tests/BUILD.bazel
@@ -175,3 +175,15 @@ ray_cc_test(
         "@com_google_googletest//:gtest_main",
     ],
 )
+
+ray_cc_test(
+    name = "throttler_test",
+    size = "small",
+    srcs = ["throttler_test.cc"],
+    tags = ["team:core"],
+    deps = [
+        "//src/ray/raylet:throttler",
+        "@com_google_absl//absl/time",
+        "@com_google_googletest//:gtest_main",
+    ],
+)
