+++ src/ray/raylet/tests/throttler_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "ray/util/throttler.h"
+#include "ray/raylet/throttler.h"
 
 #include <cstdlib>
