+++ src/ray/util/BUILD.bazel
@@ -199,14 +199,6 @@ ray_cc_library(
 )
 
 ray_cc_library(
-    name = "throttler",
-    hdrs = ["throttler.h"],
-    deps = [
-        "@com_google_absl//absl/time",
-    ],
-)
-
-ray_cc_library(
     name = "sequencer",
     hdrs = ["sequencer.h"],
     deps = [
